[SPARK-44326][SQL][CONNECT] Move utils that are used from Scala client to the common modules
### What changes were proposed in this pull request?

There are some utils are used in the scala client including ser/derse, datetime and interval utils. These can be moved to the common modules.

### Why are the changes needed?

To make sure Scala client does not depend on the Catalyst in the future.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing test

Closes #41885 from amaliujia/SPARK-44326.

Authored-by: Rui Wang <rui.wang@databricks.com>
